+++ DP.hpp
@@ -30,10 +30,10 @@
 
 #include "utils.hpp"
 #include "score2.hpp"
-#include "LTSS.hpp"
 
 #define UNUSED(expr) do { (void)(expr); } while (0)
 
+using namespace Objectives;
 using namespace IB_utils;
 
 template<typename DataType>
@@ -125,9 +125,6 @@ private:
   all_part_scores subsets_and_scores_;
   int optimal_num_clusters_OLS_;
   std::unique_ptr<ParametricContext<DataType>> context_;
-  // XXX
-  // Doesn't seem like it's needed
-  std::unique_ptr<LTSSSolver<DataType>> LTSSSolver_;
   
   void create();
   void createContext();
