+++ single_class_pmlb_driver.hpp
@@ -1,5 +1,5 @@
-#ifndef __PMLB_DRIVER_HPP__
-#define __PMLB_DRIVER_HPP__
+#ifndef __SINGLE_CLASS_PMLB_DRIVER_HPP__
+#define __SINGLE_CLASS_PMLB_DRIVER_HPP__
 
 #include <iostream>
 #include <numeric>
